# Optimization for discovering datasets for dynamic output collection.