Batch up job creation and adding datasets to the history in order to eliminate a bunch of database flushes.

Using the create_input_collection tool from PR #1958 to create an output collection with 100 elements - the optimization reduces the job finish time from 45 seconds to 15 seconds.

Before:

```
galaxy.jobs DEBUG 2016-03-21 01:50:06,659 job 23 ended (finish() executed in (46430.736 ms))
```

After:

```
galaxy.jobs DEBUG 2016-03-21 01:51:01,377 job 24 ended (finish() executed in (15474.472 ms))
```