@remimarenco 2016-03-21
Wow, 200% faster! Nice :D

About the example, you meant the other way, right?

Before:

`galaxy.jobs DEBUG 2016-03-21 01:50:06,659 job 23 ended (finish() executed in (46430.736 ms))`

After:

`galaxy.jobs DEBUG 2016-03-21 01:51:01,377 job 24 ended (finish() executed in (15474.472 ms))`